2. Bike Helmet Monitoring/Project/requirements.txt
ultralytics==8.1.0
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
Pillow==10.0.0
scipy==1.11.3
torch==2.1.0
torchvision==0.16.0
